LayoutTests/imported/w3c/web-platform-tests/custom-state-pseudo-class/idlharness.window-expected.txt


PASS idl_test setup
PASS idl_test validation
PASS Partial interface ElementInternals: original interface defined
PASS Partial interface ElementInternals: member names are unique
PASS ElementInternals includes ARIAMixin: member names are unique
FAIL CustomStateSet interface: existence and properties of interface object assert_own_property: self does not have own property "CustomStateSet" expected property "CustomStateSet" missing
FAIL CustomStateSet interface object length assert_own_property: self does not have own property "CustomStateSet" expected property "CustomStateSet" missing
FAIL CustomStateSet interface object name assert_own_property: self does not have own property "CustomStateSet" expected property "CustomStateSet" missing
FAIL CustomStateSet interface: existence and properties of interface prototype object assert_own_property: self does not have own property "CustomStateSet" expected property "CustomStateSet" missing
FAIL CustomStateSet interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "CustomStateSet" expected property "CustomStateSet" missing
FAIL CustomStateSet interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "CustomStateSet" expected property "CustomStateSet" missing
FAIL CustomStateSet interface: setlike<DOMString> undefined is not an object (evaluating 'this.get_interface_object().prototype')
FAIL CustomStateSet interface: operation add(DOMString) assert_own_property: self does not have own property "CustomStateSet" expected property "CustomStateSet" missing
FAIL CustomStateSet must be primary interface of customStateSet assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of customStateSet assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CustomStateSet interface: customStateSet must inherit property "add(DOMString)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CustomStateSet interface: calling add(DOMString) on customStateSet with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL ElementInternals interface: attribute states assert_true: The prototype object must have a property "states" expected true got false
